requirements.txt
praw
python-dotenv
scikit-learn
openai
pandas
